Fix support for requesting a custom media type
This action supports passing a custom `mediaType`, but this doesn't
actually work. The action receives a string, but
`octokit/endpoint.js` [expects][1] to be passed an object with
`format` and/or `previews` keys.

Trying to pass a `mediaType` value as suggested in the `action.yml`
leads to a `Cannot read property 'map' of undefined` error which
comes from [`endpoint.js`'s `merge` function][2] which is assuming
that the passed `mediaType` will be an object with properties.

This fixes the issue by taking the value passed in as `mediaType`
and sending it to `endpoint.js` as `mediaType.format`.

There are other potential approaches we could take to this,
for example:

* supporting a more basic `accept` option instead for a raw header
* accepting `mediaTypeFormat` and `mediaTypePreviews` (or similar)

[1]:
https://github.com/octokit/endpoint.js#endpointroute-options-or-endpointoptions
[2]:
https://github.com/octokit/endpoint.js/blob/8266fc12765e94cb966d5104a375631b4efb0dbf/src/merge.ts#L35-L37

Fixes https://github.com/octokit/request-action/issues/175.